split dynamic-form into logical chunks
