+++ src/components/common/dynamic-form/dynamic-form.js
@@ -1,363 +1,102 @@
-import { LitElement, html, css, ifDefined, nothing } from '/vendor/@lit/all@3.1.2/lit-all.min.js';
-import { serialize } from '/vendor/@shoelace/cdn@2.14.0/utilities/form.js';
-import * as i from '/components/common/dynamic-form/renders/index.js'
-import { createAlert } from '/components/common/alert.js';
-import { customElementsReady } from '/utils/custom-elements-ready.js';
-import { postConfig } from '/api/config/config.js';
-import { pkgController } from '/models/package/index.js';
+import { LitElement, html, css } from '/vendor/@lit/all@3.1.2/lit-all.min.js';
+import * as methods from '/components/common/dynamic-form/index.js';
+import { bindToClass } from '/utils/class-bind.js';
+import { styles } from './styles.js';
 
 class DynamicForm extends LitElement {
 
   static get properties() {
     return {
-      // Context
-      pupId: { type: String },
-      activeFormId: { type: String },
-
-      // Fields and values
-      fields: { type: Object },
       values: { type: Object },
-
-      // State
-      submitting: { type: Boolean },
-      dirty: { type: Boolean, reflect: true },
-
-      // Presentation
-      orientation: { type: String }
+      fields: { type: Object },
+      onSubmit: { type: Object },
+      _activeFormId: { type: String, state: true },
+      _dirty: { type: Number, state: true },
+      _loading: { type: Boolean, state: true },
+      _orientation: { type: String, reflect: true }
     };
   }
 
-  static styles = css`
-    .form-control {
-      margin-bottom: 1.5em;
-    }
-
-    .form-control.render-error sl-input::part(form-control-label),
-    .form-control.render-error sl-input::part(form-control-help-text) {
-      color: tomato;
-    }
-
-    .footer-controls {
-      display: flex;
-      justify-content: flex-end;
-    }
-
-    .footer-controls sl-button.discard-button::part(base) {
-      color: var(--sl-color-neutral-700);
-      text-decoration: underline;
-    }
-    .footer-controls sl-button.discard-button::part(base):hover {
-      color: var(--sl-color-neutral-900);
-    }
-    sl-tab.capitalize::part(base) {
-      text-transform: capitalize;
-    }
-  `;
+  static styles = styles;
 
   constructor() {
     super();
-    this.formValid = true;
-    this.activeFormId = null;
+    bindToClass(methods, this);
     this.values = {};
-    this.initialValues = {};
-    this.dirty = false;
+    this.fields = {};
+    this._activeFormId = null;
+    this._dirty = 0;
+    this._loading = false;
+    this._orientation = 'portrait';
   }
 
-  connectedCallback() {
-    super.connectedCallback();
-    this.initialValues = { ...this.values };
+  set fields(newValue) {
+    this._fields = newValue;
+    if (!newValue.sections) return;
+
+    // Create a reactive property for every form field.
+    this._initializeFormFieldProperties(newValue)
   }
 
-  getFormValues() {
-    const forms = this.shadowRoot.querySelectorAll('form');
-    let out = {}
-    forms.forEach((form) => {
-      out = {
-        ...out,
-        ...superSerialize(form)
-      }
-    })
-    return out
+  set values(newValue) {
+    if (!newValue) return;
+
+    // When an external actor (such as a  parent component, web socket) 
+    // provides new values, set them but preserve unsaved edits.
+    this._initializeValuesPreservingEdits(newValue)
   }
 
-  createFormControls(options = {}) {
-    return html`
-      <div class="footer-controls">
-        ${this.dirty ? html`
-          <sl-button
-            class="discard-button"
-            variant="text"
-            @click=${this.handleClearChanges}>
-              Discard changes
-          </sl-button>
-        ` : nothing
-        }
-        <sl-button
-          id="${options.formId}__save_button"
-          variant=primary
-          type="submit"
-          form=${options.formId}
-          ?loading=${this.submitting}
-          ?disabled=${!this.dirty}>
-        ${options.submitLabel || 'Save' }
-      </sl-button>
-      </div>
-    `
-  }
-
-  createMultipleForms(data) {
-    const tabs = data.sections.map((section, index) => {
-      const sectionId = `${(section.name || "default")}_${index}`;
-      return html`
-        <sl-tab
-          @click=${(event ) => this.handleTabChange(event, sectionId)}
-          slot="nav"
-          class="capitalize"
-          panel="${sectionId}">
-            ${section.name}
-        </sl-tab>
-      `
-    });
-
-    const form = (section, index = 0) => {
-      const sectionId = `${(section.name || "default")}_${index}`;
-      return html`
-        <form id=${sectionId}>
-          ${section.fields.map(field => this.generateField(field))}
-          ${this.createFormControls({ formId: sectionId, submitLabel: 'Save' })}
-        </form>
-      `
-    }
-
-    const panels = data.sections.map((section, index) => {
-      const sectionId = `${(section.name || "default")}_${index}`;
-      return html`
-        <sl-tab-panel name=${sectionId}>
-          ${form(section, index)}
-        </sl-tab-panel>
-      `
-    });
-
-    // if mutliple sections, render tabs and panels
-    if (data.sections.length > 1)
-    return html`
-      <sl-tab-group placement=${this.orientation === 'landscape' ? 'start' : 'top' }>
-        ${tabs}
-        ${panels}
-      </sl-tab-group>
-    `
-
-    if (data.sections.length === 1) {
-      return html`
-        ${form(data.sections[0])}
-      `
+  set _dirty(value) {
+    // Adjust only if incoming value differs, also emit change event.
+    if (this.__dirty !== value) { 
+      this.__dirty = value;
+      this._dispatchEvent('dirty-change', { dirty: this._dirty });
     }
   }
 
-  generateField(field) {
-    try {
-      if (field.hidden) return nothing;
-      return html`
-        <div class="form-control">
-          ${i[field.type](field, this.values)}
-        </div>
-      `
-    } catch (fieldRenderError) {
-      console.error('Dynamic form field error:', { field, fieldRenderError })
-      return this.generateErrorField(field)
-    } 
+  set _loading(value) {
+    // Adjust only if incoming value differs, also emit change event.
+    if (this.__loading !== value) {
+      this.__loading = value;
+      this._dispatchEvent('loading-change', { loading: this._loading });
+    }
   }
 
-  generateErrorField(field) {
-    return html`
-      <div class="form-control render-error" no-collect>
-        <sl-input
-          label="Field Error"
-          help-text="${field.type} is not a valid field type"
-          value=${field.type}
-          >
-          <sl-icon name="exclamation-diamond" slot="prefix"></sl-icon>
-        </sl-input>
-      </div>
-    `
+  get fields() {
+    return this._fields;
+  }
+
+  get values() {
+    return this._values;
+  }
+
+  get _dirty() {
+    return this.__dirty;
+  }
+
+  get _loading() {
+    return this.__loading;
   }
 
   render() {
-    if (!this.fields || !this.fields.sections) return;
-    return this.createMultipleForms(this.fields)
-  }
+    if (!this.fields?.sections) return;
 
-  handleTabChange(event, tabName) {
-    this.activeFormId = tabName;
-    this.requestUpdate();
+    return html`
+      <sl-resize-observer @sl-resize=${this._handleResize}>
+        <div class="dynamic-form-wrapper">
+          ${this._generateOneOrManyForms(this.fields)}
+        </div>
+      </sl-resize-observer>
+    `
   }
 
   async updated(changedProperties) {
-    // Check if `data` OR 'activeFormId' is the property that changed
-    const dataInitialized = changedProperties.has('fields') 
-      && (!changedProperties.has('activeFormId') || !changedProperties.activeFormId)
-
-    if (changedProperties.has('fields') || changedProperties.has('activeFormId')) {
-      // The `dataSet` has been initialized or updated OR, the activeFormId has changed, both
-      // result in some number of shoelace components being added to the DOM. 
-      
-      // Remove previous listeners
-      this._removeFormSubmitListeners();
-
-      // Determine form
-      const form = dataInitialized
-        ? this.shadowRoot.querySelector(`form`)
-        : this.shadowRoot.querySelector(`form#${this.activeFormId}`);
-
-      if (!form) {
-        return;
-      }
-
-      // Set the activeFormId when initliazing or changing data alone.
-      if (dataInitialized) { this.activeFormId = form.id }
-
-      // Wait for new elements to be ready
-      await customElementsReady(form);
-
-      // Do things knowing all custom-elements are ready.
-      this._attachFormSubmitListener(form);
-    }
-  }
-
-  handleFormChange = () => {
-    const currentValues = this.getFormValues();
-    const newDirty = !this.areValuesEqual(this.initialValues, currentValues);
-    // Check if the dirty state has changed
-    if (newDirty !== this.dirty) {
-      // Update the dirty state
-      this.dirty = newDirty;
-      // Emit the dirty-change event with the new state
-      this.dispatchEvent(new CustomEvent('dirty-change', {
-        detail: { dirty: this.dirty },
-        composed: true
-      }));
-    }
-  }
-
-  handleClearChanges() {
-    this.resetForm()
-  }
-
-  resetForm(newState) {
-    if (newState) {
-      this.initialValues = { ...this.initialValues, ...newState };
-    }
-
-    // Reset the values to their initial state
-    if (!newState) {
-      const forms = this.shadowRoot.querySelectorAll('form');
-      forms.forEach(form => form.reset());
-      this.values = { ...this.initialValues };
-    }
-
-    // Update the dirty flag
-    this.dirty = false;
-    this.dispatchEvent(new CustomEvent('dirty-change', {
-      detail: { dirty: this.dirty },
-      composed: true
-    }));
-  }
-
-  areValuesEqual(objectA, objectB) {
-    if (!objectA || !objectB) return false;
-    const keysA = Object.keys(objectA);
-    const keysB = Object.keys(objectB);
-
-    // Check if both objects have the same number of properties
-    if (keysA.length !== keysB.length) {
-      return false;
-    }
-
-    // Check if values for each property are the same
-    for (const key of keysA) {
-      if (objectA[key] !== objectB[key]) {
-        return false;
-      }
-    }
-
-    return true;
-  }
-
-  _formSubmitHandler = async (event) => {
-    event.preventDefault();
-    this.submitting = true;
-    try {
-      // Collect data
-      const form = this.shadowRoot.querySelector(`form#${this.activeFormId}`);
-      const data = superSerialize(form);
-
-      // Submit data
-      const res = await postConfig(this.pupId, data);
-
-      // Update local model.
-      pkgController.savePupChanges(this.pupId, data);
-
-      // Celebrate successs.
-      this.submitting = false;
-      const saveButton = this.shadowRoot.querySelector(`form#${this.activeFormId} sl-button[type="submit"]`);
-      const buttonOriginalText = saveButton.textContent;
-      saveButton.textContent = 'Done';
-      saveButton.disabled = true;
-
-      setTimeout(function () {
-        saveButton.textContent = buttonOriginalText;
-        saveButton.disabled = false;
-      }, 1400);
-
-      // Reset form
-      this.resetForm(data);
-
-    } catch (error) {
-
-      // Display errors
-      createAlert('danger', `Error: ${error.message}`);
-      console.error('Submission error', error);
-
-    }
-  };
-
-  _attachFormSubmitListener(form) {
-    form.addEventListener('submit', this._formSubmitHandler);
-    form.addEventListener('sl-change', this.handleFormChange);
-  }
-
-  _removeFormSubmitListeners() {
-    // Remove event listeners from all forms.
-    const forms = this.shadowRoot.querySelectorAll('form');
-    forms.forEach((form) => {
-      form.removeEventListener('submit', this._formSubmitHandler);
-      form.removeEventListener('sl-change', this.handleFormChange);
-    });
+    await this._onUpdate(changedProperties);
   }
 
   disconnectedCallback() {
     this._removeFormSubmitListeners();
   }
-}
-
-// superSerialize performs one extra step after form serialization.
-// when a user toggles a checkbox from checked to unchecked, shoelace serialize
-// does not include that form field within its serialized object.
-// Therefore, superSerialize inspects all toggles and checkboxes and
-// explicitly sets their state as "on" or "off".
-function superSerialize(form) {
-  let data = serialize(form)
-  const toggles = form.querySelectorAll('sl-switch') || [];
-  const checkboxes = form.querySelectorAll('sl-checkbox') || [];
-
-  [...toggles, ...checkboxes].forEach(tog => {
-    if (!data[tog.name]) {
-      const value = tog.checked ? 'on' : 'off';
-      data[tog.name] = value
-    }
-  });
-
-  return data
 }
 
 customElements.define('dynamic-form', DynamicForm);
