+++ src/components/common/dynamic-form/index.js
@@ -1,0 +1,2 @@
+export * from './lib/index.js';
+export * from './renders/index.js';
