+++ src/components/common/dynamic-form/lib/events.js
@@ -1,0 +1,8 @@
+export function _dispatchEvent(name, detail) {
+  this.dispatchEvent(new CustomEvent(
+  `form-${name}`, {
+    detail,
+    composed: true,
+    bubbles: true,
+  }));
+}
