+++ src/components/common/dynamic-form/lib/on-update.js
@@ -1,0 +1,51 @@
+import { customElementsReady } from '/utils/custom-elements-ready.js';
+
+export async function _onUpdate(changedProperties) {
+  if (!this._shouldUpdateForm(changedProperties)) {
+    return;
+  }
+
+  // Remove previous form submit listeners
+  this._removeFormSubmitListeners();
+
+  // Determine the appropriate form to target
+  const form = this._getTargetForm(changedProperties);
+  if (!form) {
+    console.log("No form found.");
+    return;
+  }
+
+  // Update the active form ID if necessary
+  this._updateActiveFormId(form, changedProperties);
+
+  // Ensure all custom elements within the form are fully defined
+  await customElementsReady(form);
+
+  // Attach a new submit listener to the form
+  this._attachFormSubmitListener(form);
+}
+
+export function _shouldUpdateForm(changedProperties) {
+  return changedProperties.has('fields') || changedProperties.has('_activeFormId');
+}
+
+export function _getTargetForm(changedProperties) {
+    const isDataInitialization =
+      changedProperties.has('fields') &&
+      (!changedProperties.has('_activeFormId') || !this._activeFormId);
+
+    const formSelector = isDataInitialization
+      ? 'form'
+      : `form#${this._activeFormId}`;
+
+  return this.shadowRoot.querySelector(formSelector);
+}
+
+export function _updateActiveFormId(form, changedProperties) {
+  const isDataInitialization = changedProperties.has('fields') &&
+      (!changedProperties.has('_activeFormId') || !this._activeFormId);
+
+  if (isDataInitialization) {
+      this._activeFormId = form.id;
+  }
+}
