+++ src/components/common/dynamic-form/renders/fields/text-number.js
@@ -2,26 +2,31 @@ import { html, ifDefined } from '/vendor/@lit/all@3.1.2/lit-all.min.js';
 
 const ifd = ifDefined
 
-export function _render_password(field) {
+export function _render_number(field) {
   return html`
     <sl-input
-      type="password"
+      type="number"
       name=${field.name}
       label=${ifd(field.label)}
       placeholder=${ifd(field.placeholder)}
       help-text=${ifd(field.help)}
-      minlength=${ifd(field.minLength)}
-      maxlength=${ifd(field.maxLength)}
-      pattern=${ifd(field.pattern)}
+      min=${ifd(field.min)}
+      max=${ifd(field.max)}
+      step=${ifd(field.step)}
       size=${ifd(field.size)}
       .value=${this[field.name]}
       ?clearable=${field.clearable}
+      ?noSpinButtons=${field.noSpinButtons}
+      ?autofocus=${field.autofocus}
       ?required=${field.required}
       ?disabled=${field.disabled}
-      ?password-toggle=${field.passwordToggle}
       ?data-dirty-field=${this[this._dirtyFlagField(field.name)]}
-      @input=${this.handleInput}
+      @input=${this._handleInput}
       >
     </sl-input>
   `;
-} 
+}
+
+function valueParser(value) {
+  return value.toString()
+}
