+++ src/components/common/dynamic-form/renders/index.js
@@ -1,15 +1,15 @@
-export * from './checkbox.js';
-export * from './color.js';
-export * from './date.js';
-export * from './index.js';
-export * from './radio-button.js';
-export * from './radio.js';
-export * from './range.js';
-export * from './rating.js';
-export * from './select.js';
-export * from './text-email.js';
-export * from './text-number.js';
-export * from './text-password.js';
-export * from './text-string.js';
-export * from './textarea.js';
-export * from './toggle.js';
+export * from './fields/checkbox.js';
+export * from './fields/color.js';
+export * from './fields/date.js';
+export * from './fields/radio-button.js';
+export * from './fields/radio.js';
+export * from './fields/range.js';
+export * from './fields/rating.js';
+export * from './fields/select.js';
+export * from './fields/text-email.js';
+export * from './fields/text-number.js';
+export * from './fields/text-password.js';
+export * from './fields/text-string.js';
+export * from './fields/textarea.js';
+export * from './fields/toggle.js';
+export * from './form.js';
