+++ src/components/views/pup-snapshot/pup-snapshot.js
@@ -2,7 +2,7 @@ import { LitElement, html, nothing } from '/vendor/@lit/all@3.1.2/lit-all.min.js
 import { bindToClass } from '/utils/class-bind.js'
 import { asyncTimeout } from '/utils/timeout.js'
 import { styles } from './pup-snapshot.styles.js';
-import '/components/common/dynamic-form/dynamic-form-reuse.js'
+import '/components/common/dynamic-form/dynamic-form.js'
 import '/components/common/animated-dots.js'
 import '/components/common/sparkline-chart/sparkline-chart.js'
 import '/components/views/log-viewer/log-viewer.js'
